+++ packages/magic-search/src/components/PlaceholderPrompt.tsx
@@ -9,7 +9,7 @@ const PlaceholderPrompt = () => {
   return (
     <div
       className={twMerge(
-        `py-4 px-6 flex flex-col gap-2 bg-white rounded-3xl mb-3 ${getClassOverride(PROMPT_ID, configuration)}`,
+        `py-4 px-5 flex flex-col gap-2 bg-white rounded-3xl mb-3 ${getClassOverride(PROMPT_ID, configuration)}`,
       )}
     >
       <Shimmer width={300} />